backend/requirements.txt
fastapi
uvicorn[standard]
qdrant-client
sentence-transformers
nltk
pydantic-settings
python-dotenv
httpx
umap-learn
scikit-learn
